requirements.txt
# Core Dependencies
streamlit==1.37.1
groq==0.11.0
python-dotenv==1.0.1

# Data Processing & Visualization
matplotlib==3.9.0
pandas==2.2.2
plotly==5.17.0
seaborn==0.13.0

# Document Generation & Processing
reportlab==4.2.2
fpdf2==2.7.9
python-docx==1.1.0
openpyxl==3.1.2

# Database & Analytics
sqlite3  # Built-in with Python
sqlalchemy==2.0.23
textblob==0.17.1
scikit-learn==1.3.2

# API & Web Development
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
bcrypt==4.1.2
passlib==1.7.4
python-jose==3.3.0

# Utility Libraries
pathlib  # Built-in with Python
typing-extensions==4.8.0
pydantic==2.5.0
click==8.1.7

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Optional Advanced Features
redis==5.0.1
celery==5.3.4
pillow==10.1.0
python-multipart==0.0.6
plotly==5.17.0
numpy==1.24.3
sqlalchemy==2.0.21
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
reportlab==4.0.4